符号数计算混用修改

+++ frameworks/native/request/src/request_service_proxy.cpp
@@ -767,9 +767,9 @@ int32_t RequestServiceProxy::OpenChannel(int32_t &sockFd)
         return errCode;
     }
     sockFd = reply.ReadFileDescriptor();
-    int flags = fcntl(sockFd, F_GETFL, 0);
+    uint32_t flags = static_cast<uint32_t>(fcntl(sockFd, F_GETFL, 0));
     fcntl(sockFd, F_SETFL, flags & ~O_NONBLOCK);
-    flags = fcntl(sockFd, F_GETFL, 0);
+    flags = static_cast<uint32_t>(fcntl(sockFd, F_GETFL, 0));
     if (flags & O_NONBLOCK) {
         REQUEST_HILOGE("Set ~O_NONBLOCK failed for fd %{public}d", sockFd);
     }
